backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
Flask-SocketIO>=5.3.6
simple-websocket>=1.0.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
cachetools>=5.3.0
pytz>=2024.1
websocket-client>=1.6.0
python-dotenv>=1.0.1
pymongo>=4.6.0
flask-bcrypt>=1.0.1